created requriements.txt
Bot setup is unnecessarily complicated, I added a requirements.txt file to handle dependencies for the user, now user only needs to type one command to install all dependencies. I didn't test it on any other systems to see if it works or not, my setup is a macbook pro with visual studio code instead of windows and pycharm, so you should verify it works. Other improvements that need to be made to speed up setup include:
1) create config.py file with object that contains all the user specific configuration like CVV number, firefox profile path etc.
2) update instructions for using terminal only or other IDE and not specifically just pycharm